requirements.txt
configparser==4.0.2
Django==1.11.21
django-markdown-editor==1.0.3
django-mdeditor==0.1.14
Jinja2==2.7.2
jsonpatch==1.2
jsonpointer==1.9
Markdown==3.1.1
Markdown-Editor==1.0.3
markdown2==2.3.8
mysql-connector-python==8.0.16
#MySQL-python==1.2.5
paramiko==2.5.0
requests==2.6.0
urllib3==1.10.2
#uWSGI==2.0.18
mysqlclient
djangorestframework-3.11.0